Update 05.the_main_game_scene.rst
I followed the guide to create the first game with Godot, and i found it confusing that Mobs would not spawn following the MobSpawnLocation. 
After some debugging i figured it was because they spawn at local position and not global position.
To make it more clear to newcomers i suggest either using GlobalPosition OR adding a note that explains the difference between local and global position. I opted for the first solution and created this pull request with the change.